requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
pathlib==1.0.1
pathlib2==2.3.5
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
